fix: Pull personal_allowance into its own file and clarify usage
